refactor: :fire: refactored create token components
