+++ src/features/Apiexplorer/styles.module.scss
@@ -11,7 +11,6 @@
   justify-content: space-between;
   width: 100%;
   margin-top: rem(4);
-  min-height: 89vh;
 
   .heading {
     text-align: center;
@@ -20,11 +19,6 @@
   .pageWrapper {
     display: flex;
     color: var(--ifm-color-emphasis-1000);
-
-    @media screen and (min-width: 1440px) {
-      width: 1310px;
-      margin: auto;
-    }
 
     .playground {
       display: flex;
