+++ src/features/dashboard/components/token-register/token-register.tsx
@@ -6,9 +6,9 @@ import { scopesObjectToArray } from '@site/src/utils';
 import ApiTokenCard from '../api-token-card';
 import useCreateToken from '@site/src/features/dashboard/hooks/useCreateToken';
 import * as yup from 'yup';
-import CreateTokenField from '../api-token-form';
-import AccountSwitcher from '@site/src/components/AccountSwitcher';
 import './token-register.scss';
+import CreateTokenField from '../api-token-form/create-token-field';
+import AccountSwitcher from '@site/src/components/AccountSwitcher';
 
 const schema = yup
   .object({
@@ -86,8 +86,6 @@ const TokenRegister = (props: HTMLAttributes<HTMLFormElement>) => {
   const {
     handleSubmit,
     register,
-    setValue,
-    getValues,
     reset,
     formState: { errors },
   } = useForm<TApiTokenForm>({
