+++ src/styles/index.scss
@@ -280,7 +280,7 @@ div[class*='sidebarViewport'] {
 }
 
 .navbar {
-  padding: 0 5vw;
+  padding: 0;
   height: var(--nav-height);
   line-height: rem(2);
   @media (max-width: 996px) {
@@ -290,121 +290,9 @@ div[class*='sidebarViewport'] {
     display: flex;
     align-items: center;
     gap: rem(1);
-    .search-overlay {
-      display: none;
-    }
-    &.search-closed ~ div[class*='navbarSearch'] {
-      display: none;
-    }
-    &.search-open {
-      .search-overlay {
-        display: inline-block;
-        position: fixed;
-        width: 100vw;
-        height: 100vh;
-        background-color: rgba(0, 0, 0, 0.4);
-        inset: 0;
-
-        @media (max-width: 768px) {
-          width: rem(2.5);
-          height: rem(2.5);
-          background-color: unset;
-          background-image: url('/img/close_dialog.svg');
-          background-repeat: no-repeat;
-          background-position: center;
-          background-size: rem(1.5);
-          inset: rem(1) rem(1) auto auto;
-          z-index: 1;
-        }
-      }
-      & + div[class*='searchBox'] {
-        position: fixed;
-        min-width: rem(32);
-        max-width: rem(79.2);
-        max-height: rem(48);
-        top: 5%;
-        width: 100%;
-        height: 100%;
-        border-radius: rem(1.6);
-        left: 50%;
-        transform: translateX(-50%);
-        padding: 0;
-        background-color: var(--ifm-color-white);
-
-        @media (max-width: 768px) {
-          min-width: unset;
-          max-width: unset;
-          max-height: unset;
-          border-radius: unset;
-          top: 0;
-        }
-
-        span[class*='searchBar'] {
-          width: 100%;
-          span[class*='dropdownMenu'] {
-            width: 100%;
-            max-width: unset;
-            background-color: unset;
-            box-shadow: none;
-            padding: 0;
-            div[class*='hitFooter'] a {
-              color: var(--colors-coral500);
-            }
-            div[class*='suggestion'] {
-              box-shadow: none;
-              border-radius: 0;
-              height: rem(4);
-              span[class*='hitAction'] {
-                display: none;
-              }
-              span[class*='hitWrapper'] {
-                span[class*='hitTitle'],
-                span[class*='hitPath'] {
-                  text-align: left;
-                  mark {
-                    color: var(--colors-coral500);
-                  }
-                }
-              }
-              &[class*='cursor'] {
-                background-color: var(--search-result);
-              }
-            }
-          }
-        }
-        .navbar__search {
-          margin: 0;
-          margin-top: rem(2);
-          display: flex;
-          justify-content: center;
-          text-align: center;
-          width: 100%;
-
-          @media (max-width: 768px) {
-            margin-top: rem(4);
-          }
-
-          button[class*='searchClearButton'] {
-            right: rem(3);
-          }
-        }
-        div[class*='searchHintContainer'] {
-          right: 5%;
-        }
-      }
-    }
-  }
-  &__search-input {
-    background: var(--ifm-color-white) url(/img/search.svg) rem(1.2) center / rem(1.6) rem(1.6);
-    background-repeat: no-repeat;
-    border: 1px solid var(--colors-greyLight500);
-    border-radius: rem(1.5);
-    height: rem(4);
-    width: calc(100% - 1.25rem) !important;
   }
   &__inner {
     @media screen and (min-width: 1440px) {
-      max-width: 1310px;
       gap: rem(14);
       margin: auto;
     }
